Refactor search functionality: replace 'search' with 'lookup' in comparison modules and remove deprecated search.py
