+++ libs/utils/dictutil.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Any, Protocol
 import libs.global_value as g
 from libs.domain.command import CommandParser
 from libs.domain.placeholder import PlaceholderBuilder
-from libs.functions import lookup, search
+from libs.functions import lookup
 from libs.types import ChannelType, StyleOptions
 from libs.utils import formatter
 from libs.utils.timekit import ExtendedDatetime as ExtDt
@@ -178,7 +178,7 @@ def placeholder(subcom: "SubCommandLike", m: "MessageParserProtocol") -> Placeho
             params.stipulated = 1
 
     if departure_time.range(search_range).start == ExtDt("1900-01-01 00:00:00.000000"):
-        params.starttime = search.first_record(
+        params.starttime = lookup.first_record(
             g.cfg.rule.get_version(
                 mode=params.mode,
                 mapping=not (params.mixed),
